# apps/server/swagger-spec.json
{"openapi":"3.0.0","paths":{"/user/{id}":{"get":{"operationId":"UserController_findUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["user"]}},"/auth/register":{"post":{"operationId":"AuthController_register","summary":"User Registration","parameters":[{"name":"isCookie","required":false,"in":"query","description":"true/false","schema":{"type":"string"}},{"name":"isOauth","required":false,"in":"query","description":"true/false","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"User Registration Details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"default":{"description":"User Registration Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"tags":["Authentication"]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"User Login","parameters":[{"name":"isCookie","required":false,"in":"query","description":"true/false","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"User Login Details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/auth/refresh-token":{"get":{"operationId":"AuthController_refreshToken","summary":"RefreshTokens","parameters":[{"name":"rt","required":false,"in":"query","description":"Refresh Token","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/auth/request-reset-password":{"get":{"operationId":"AuthController_requestResetPassword","summary":"Request Password Reset","parameters":[{"name":"email","required":true,"in":"query","description":"Email Address","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Authentication"]}},"/auth/change-password":{"post":{"operationId":"AuthController_changePassword","summary":"Change Password","parameters":[],"requestBody":{"required":true,"description":"New Password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["Authentication"]}}},"info":{"title":"Nestjs API","description":"The nestjs API description","version":"1.5","contact":{}},"tags":[],"servers":[{"url":"http://localhost:4000/","description":"Local environment"},{"url":"https://staging.yourapi.com/","description":"Staging"},{"url":"https://production.yourapi.com/","description":"Production"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserResponseDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"disable_access":{"type":"boolean"},"email_verified":{"type":"boolean"},"first_name":{"type":"string"},"is2fa":{"type":"boolean"},"last_name":{"type":"string"},"otp_secret":{"type":"string"},"phone":{"type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string","description":"Email"},"password":{"type":"string","description":"Password"}},"required":["email","password"]},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"first_name":{"type":"string"},"last_name":{"type":"string"},"email_verified":{"type":"boolean"},"is2fa":{"type":"boolean"},"disable_access":{"type":"boolean"},"otp_secret":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","email","phone","first_name","last_name","email_verified","is2fa","disable_access","otp_secret","created_at","updated_at"]},"TokensDto":{"type":"object","properties":{"accessToken":{"type":"string","description":"Access Token"},"refreshToken":{"type":"string","description":"Refresh Token"}},"required":["accessToken","refreshToken"]},"LoginResponseDto":{"type":"object","properties":{"user":{"description":"Token Type","allOf":[{"$ref":"#/components/schemas/UserDto"}]},"tokens":{"description":"Token Type","allOf":[{"$ref":"#/components/schemas/TokensDto"}]}},"required":["user","tokens"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"Email"},"password":{"type":"string","description":"Password"}},"required":["email","password"]},"ChangePasswordDto":{"type":"object","properties":{"email":{"type":"string","description":"Email"},"otp":{"type":"string","description":"OTP"},"newPassword":{"type":"string","description":"New Password"}},"required":["email","otp","newPassword"]}}}}